Move the predicates to the basic launch view
